+++ test/TestPrometheusMiddleware.AspNetCore.2.1/Startup.cs
@@ -1,0 +1,61 @@
+﻿// <copyright file="Startup.cs" company="OpenCensus Authors">
+// Copyright 2018, OpenCensus Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using OpenCensus.Stats;
+using OpenCensus.Tags;
+using System.Net.Http;
+using OpenCensus.Exporter.Prometheus.Middleware;
+using OpenCensus.Exporter.Prometheus;
+
+
+namespace TestPrometheusMiddleware.AspNetCore._2._0
+{
+    public class Startup
+    {
+        public Startup(IConfiguration configuration)
+        {
+            Configuration = configuration;
+        }
+
+        public IConfiguration Configuration { get; }
+
+        // This method gets called by the runtime. Use this method to add services to the container.
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddMvc();
+            services.AddSingleton<IViewManager>(Stats.ViewManager);
+            services.AddSingleton<IStatsRecorder>(Stats.StatsRecorder);
+            services.AddSingleton<ITagger>(Tags.Tagger);
+        }
+
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+
+            app.UsePrometheusExporter(new PrometheusMiddlewareOptions() { Path = "/metrics" }, Stats.ViewManager);
+            app.UseMvc();
+        }
+    }
+}